backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["vadim-shalapugin <vadimsalapugin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django-rest-framework = "^0.1.0"
python-dotenv = "^0.21.0"
django-cors-headers = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
